\name{mass.spec}
\alias{mass.spec}
\docType{data}
\title{
 Mass spectrometry
}
\description{
  Five samples of mass spec data.
}
\usage{data("mass.spec")}
\format{
  List of five data frames that can be used with PeakSegFPOPchrom or
  PeakSegPDPAchrom.
}
\source{
  Pier-Luc Plante
}
\keyword{datasets}
