\name{McGill0003_H3K4me3_chr1}
\alias{McGill0003_H3K4me3_chr1}
\docType{data}
\title{
  Whole chromosome ChIP-seq
}
\description{
  All of the data on chr1 from one ChIP-seq experiment.
}
\usage{data("McGill0003_H3K4me3_chr1")}
\format{
  A data frame with 13623262 observations on the following 3 variables.
  \describe{
    \item{\code{chromStart}}{a numeric vector}
    \item{\code{chromEnd}}{a numeric vector}
    \item{\code{count}}{a numeric vector}
  }
}
\source{
  ~/projects/chip-seq/data/McGill0003.1.brainTumor.H3K4me1.coverage.bedGraph
}
\keyword{datasets}
